+++ zest/releaser/pypi.py
@@ -304,6 +304,31 @@ class PypiConfig(BaseConfig):
             return default
         return result
 
+    def encoding(self):
+        """Return encoding to use for text files.
+
+        Mostly the changelog and if needed `setup.py`.
+
+        The encoding cannot always be determined correctly.
+        This setting is a way to fix that.
+        See https://github.com/zestsoftware/zest.releaser/issues/264
+
+        Configure this by adding an ``encoding`` option, either in the
+        package you want to release, or in your ~/.pypirc::
+
+            [zest.releaser]
+            encoding = utf-8
+        """
+        default = ''
+        if self.config is None:
+            return default
+        try:
+            result = self._get_text(
+                'zest.releaser', 'encoding', default=default, raw=True)
+        except (NoSectionError, NoOptionError, ValueError):
+            return default
+        return result
+
     def create_wheel(self):
         """Should we create a Python wheel for this package?